VSL-52: Reorganized deployment code, added deployment of contract so tests work
