+++ contracts/core/FiatToken.cdc
@@ -1,6 +1,6 @@
 import Crypto
-import FungibleToken from 0x{{.FungibleToken}}
-import OnChainMultiSig from 0x{{.OnChainMultiSig}}
+import FungibleToken from 0xee82856bf20e2aa6
+import OnChainMultiSig from 0xf8d6e0586b0a20c7
 
 pub contract FiatToken: FungibleToken {
