+++ flow.json
@@ -60,12 +60,14 @@
 		"OnChainMultiSig": {
 			"source": "./contracts/core/OnChainMultiSig.cdc",
 			"aliases": {
+				"emulator": "0xf8d6e0586b0a20c7",
 				"testnet": "0xa983fecbed621163"
 			}
 		},
 		"FiatToken": {
 			"source": "./contracts/core/FiatToken.cdc",
 			"aliases": {
+				"emulator": "0xf8d6e0586b0a20c7",
 				"testnet": "0xa983fecbed621163"
 			}
 		}
