+++ test_utils.go
@@ -8,6 +8,7 @@ import (
 	"strconv"
 	"testing"
 
+	"github.com/DapperCollectives/VESSEL/go/util"
 	"github.com/bjartek/overflow/overflow"
 	"github.com/onflow/cadence"
 	"github.com/onflow/flow-go-sdk"
@@ -21,8 +22,10 @@ type OverflowTestUtils struct {
 const ServiceAddress = "0xf8d6e0586b0a20c7"
 
 func NewOverflowTest(t *testing.T) *OverflowTestUtils {
-	return &OverflowTestUtils{T: t, O: overflow.NewTestingEmulator().Start()}
-	// return &OverflowTestUtils{T: t, O: overflow.NewOverflowEmulator().Start()}
+	//otu := &OverflowTestUtils{T: t, O: overflow.NewTestingEmulator().Start()}
+	otu := &OverflowTestUtils{T: t, O: overflow.NewOverflowEmulator().Start()}
+	util.DeployFiatToken("treasuryOwner", "USDC", "0.1.0", otu.O)
+	return otu
 }
 
 func (otu *OverflowTestUtils) SetupTreasury(name string, signers []string, threshold uint64) *OverflowTestUtils {
